Struct tartan_elf::ProgramHeader32

#[repr(C)]
pub struct ProgramHeader32 {
    pub segment_type: ProgramSegmentType,
    pub file_offset: u32,
    pub virtual_addr: u32,
    pub physical_addr: u32,
    pub file_size: u32,
    pub mem_size: u32,
    pub flags: ProgramSegmentFlags,
    pub alignment: u32,
}

Header that controls loading of a single segment of an executable (32-bit variant)

Fields
segment_type: ProgramSegmentType

Indicates the contents/meaning of the segment described by this header.

file_offset: u32

File offset to the start of this segment’s data.

virtual_addr: u32

Virtual base address where the segment should be loaded.

physical_addr: u32

Physical base address where the segment should be loaded. Typically ignored.

file_size: u32

Size in bytes of the file data for this segment.

mem_size: u32

Size in bytes of this segment once it is loaded into memory.

flags: ProgramSegmentFlags

Controls permissions and other OS/arch-specific flags.

alignment: u32

Alignment of the segment in bytes, for both the file and memory.