Struct tartan_elf::SectionFlags

#[repr(transparent)]
pub struct SectionFlags(_);

Additional attributes for object file sections.

Implementations

impl SectionFlags

pub fn write(&self) -> bool

Indicates the section contains data that should be writable by the process.

pub fn allocate(&self) -> bool

Indicates this section should be loaded into memory when the process is run.

pub fn execute(&self) -> bool

Indicates this section contains code that should be executable by the process.

pub fn merged(&self) -> bool

Indicates that multiple sections of this type can be combined while deleting duplicate contents.

pub fn strings(&self) -> bool

Indicates the section contains null-terminate strings.

pub fn extra_info_is_link(&self) -> bool

Indicates that the extra_info field of the section header contains an index to another section header.

pub fn keep_order(&self) -> bool

Indicates that this section and the [associated_section_index] should maintain their relative order when linking.

pub fn os_nonconforming(&self) -> bool

Indicates that the section must be processed in an OS-defined way during linking.

pub fn group_member(&self) -> bool

Indicates the section is listed by a section with type SectionType::Group.

pub fn thread_locals(&self) -> bool

Indicates the section contains a template for the thread-local storage area.

pub fn os_defined(&self) -> u8

Bits that are left for OS-defined flags.

pub fn arch_defined(&self) -> u8

Bits that are left for CPU architecture-defined flags.

pub fn set_write(&mut self, value: bool)

Indicates the section contains data that should be writable by the process.

pub fn set_allocate(&mut self, value: bool)

Indicates this section should be loaded into memory when the process is run.

pub fn set_execute(&mut self, value: bool)

Indicates this section contains code that should be executable by the process.

pub fn set_merged(&mut self, value: bool)

Indicates that multiple sections of this type can be combined while deleting duplicate contents.

pub fn set_strings(&mut self, value: bool)

Indicates the section contains null-terminate strings.

pub fn set_extra_info_is_link(&mut self, value: bool)

Indicates that the extra_info field of the section header contains an index to another section header.

pub fn set_keep_order(&mut self, value: bool)

Indicates that this section and the [associated_section_index] should maintain their relative order when linking.

pub fn set_os_nonconforming(&mut self, value: bool)

Indicates that the section must be processed in an OS-defined way during linking.

pub fn set_group_member(&mut self, value: bool)

Indicates the section is listed by a section with type SectionType::Group.

pub fn set_thread_locals(&mut self, value: bool)

Indicates the section contains a template for the thread-local storage area.

pub fn set_os_defined(&mut self, value: u8)

Bits that are left for OS-defined flags.

pub fn set_arch_defined(&mut self, value: u8)

Bits that are left for CPU architecture-defined flags.