Type Alias tartan_elf::ProgramHeaderNative

pub type ProgramHeaderNative = ProgramHeader64;

Program header variant native to the current target

Aliased Type

struct ProgramHeaderNative {
    pub segment_type: ProgramSegmentType,
    pub flags: ProgramSegmentFlags,
    pub file_offset: u64,
    pub virtual_addr: u64,
    pub physical_addr: u64,
    pub file_size: u64,
    pub mem_size: u64,
    pub alignment: u64,
}

Fields
segment_type: ProgramSegmentType

Indicates the contents/meaning of the segment described by this header.

flags: ProgramSegmentFlags

Controls permissions and other OS/arch-specific flags.

file_offset: u64

File offset to the start of this segment’s data.

virtual_addr: u64

Virtual base address where the segment should be loaded.

physical_addr: u64

Physical base address where the segment should be loaded. Typically ignored.

file_size: u64

Size in bytes of the file data for this segment.

mem_size: u64

Size in bytes of this segment once it is loaded into memory.

alignment: u64

Alignment of the segment in bytes, for both the file and memory.