Crate tartan_elf

Support for Executable and Linkable Format (ELF) binaries.

Structs

Endianness
Specifies whether the file uses big or little endian byte ordering.

FileClass
Specifies the address size (32- or 64-bit) of the file.

FileType
Types of object files supported by the ELF specification

Header
File header for an ELF file with a specific address size.

HeaderIdent
Initial header which has the same layout in all ELF variants and determines how to interpret the rest of the file (endianness, sizes).

Machine
Identifies the target CPU architecture.

OSABI
Operating-system-specific ABI extensions supported by the ELF specification.

ProgramHeader32
Header that controls loading of a single segment of an executable (32-bit variant)

ProgramHeader64
Header that controls loading of a single segment of an executable (64-bit variant)

ProgramSegmentFlags
Permissions and other OS/arch-specific flags for individual program segments.

ProgramSegmentType
Defines the contents and meaning of individual program segments.

SectionFlags
Additional attributes for object file sections.

SectionHeader
Header that controls linking

SectionType
Defines the contents and meaning of individual sections.

Type Definitions

HeaderNative
Variant of the Header structure native to the current target.

ProgramHeaderNative
Program header variant native to the current target

SectionHeaderNative
Variant of the SectionHeader structure native to the current target